requirements.txt
biopython
pandas
spacy
scispacy
transformers
torch
psycopg2-binary
fastapi
uvicorn
streamlit
